Updated Germ versions. Automatic deletion of temp-files in HTTP requests.

+++ lib/hayabusa_http_session_request.rb
@@ -240,10 +240,10 @@ class Hayabusa::Http_session::Request
     end
   end
   
-  #Deletes all tempfiles created by this object.
+  #Deletes all tempfiles created by this object. This is useually called from Http_session at the end of a request.
   def delete_tempfiles
     @files_arr.delete_if do |tempfile_path|
-      File.unlink(tempfile_path)
+      File.unlink(tempfile_path) if File.exists?(tempfile_path)
       true
     end
   end
